Create Motor Button Control
Pin 9 is the servo output for the motor.
Pin 7/8 are button inputs where 7 will increase the speed of the motor and 8 will decrease the speed.
There is a fail-safe for both the upper and lower limits of when the motor will turn on ... based on trial and error, though I didn't have the propeller attached. These values may change.